syncing server

+++ train.py
@@ -256,7 +256,7 @@ def train_loop(train_params, model_params):
                                 #This represents the expected log_likelihood corresponding to each target y_i in the mini batch
 
                         kl_loss_weight = utility.kl_loss_weighting_scheme(train_set_size_batches) #TODO: Implement scheme where kl loss increases during training
-                        kl_loss = tf.math.reduce_sum( model.losses ) * kl_loss_weight * (1/model_params['model_type_settings']['stochastic_f_pass'])  #This KL-loss is already normalized against the number of samples of weights drawn #TODO: Later implement your own Adam type method to determine this
+                        kl_loss = tf.cast( tf.math.reduce_sum( model.losses ) * kl_loss_weight * (1/model_params['model_type_settings']['stochastic_f_pass']), tf.float32)  #This KL-loss is already normalized against the number of samples of weights drawn #TODO: Later implement your own Adam type method to determine this
                         
                         var_free_nrg_loss = kl_loss  - log_likelihood
